+++ mne_bids/tests/test_utils.py
@@ -59,10 +59,8 @@ def return_bids_test_dir(tmpdir_factory):
 
     raw = mne.io.read_raw_fif(raw_fname)
     # Write multiple runs for test_purposes
-    bids_basename2 = bids_basename.replace('run-{}'.format(run), 'run-02')
-    for name in [bids_basename,
-                 bids_basename2,
-                 ]:
+    for run_idx in [run, '02']:
+        name = bids_basename.copy().update(run=run_idx)
         with pytest.warns(RuntimeWarning, match='No line frequency'):
             write_raw_bids(raw, name, bids_root,
                            events_data=events_fname, event_id=event_id,
@@ -180,7 +178,7 @@ def test_make_filenames():
     prefix_data = dict(subject='one', session='two', task='three',
                        acquisition='four', run='five', processing='six',
                        recording='seven', suffix='suffix.csv')
-    assert make_bids_basename(**prefix_data) == 'sub-one_ses-two_task-three_acq-four_run-five_proc-six_recording-seven_suffix.csv'  # noqa
+    assert str(make_bids_basename(**prefix_data)) == 'sub-one_ses-two_task-three_acq-four_run-five_proc-six_rec-seven_suffix.csv'  # noqa
 
     # subsets of keys works
     assert make_bids_basename(subject='one', task='three', run=4) == 'sub-one_task-three_run-04'  # noqa
@@ -287,7 +285,7 @@ def test_parse_bids_filename(fname):
     assert params['task'] == 'test'
     assert params['split'] == '01'
     assert list(params.keys()) == ['sub', 'ses', 'task', 'acq', 'run', 'proc',
-                                   'space', 'recording', 'split', 'kind']
+                                   'space', 'rec', 'split', 'kind']
 
 
 def test_age_on_date():
@@ -370,3 +368,61 @@ def test_find_matching_sidecar(return_bids_test_dir):
     # Find nothing but receive None, because we set `allow_fail` to True
     with pytest.warns(RuntimeWarning, match='Did not find any'):
         _find_matching_sidecar(bids_basename, bids_root, 'foo.bogus', True)
+
+
+def test_bids_path(return_bids_test_dir):
+    """Test usage of BIDSPath object."""
+    bids_root = return_bids_test_dir
+
+    bids_basename = make_bids_basename(
+        subject=subject_id, session=session_id, run=run, acquisition=acq,
+        task=task)
+
+    # get_bids_fname should fire warning
+    with pytest.raises(ValueError, match='No filename extension was provided'):
+        bids_fname = bids_basename.get_bids_fname()
+
+    # should find the correct filename if bids_root was passed
+    bids_fname = bids_basename.get_bids_fname(bids_root=bids_root)
+    assert bids_fname == bids_basename.update(suffix='meg.fif')
+
+    # confirm BIDSPath assigns properties correctly
+    bids_basename = make_bids_basename(subject=subject_id,
+                                       session=session_id)
+    assert bids_basename.subject == subject_id
+    assert bids_basename.session == session_id
+    assert 'subject' in bids_basename.entities
+    assert 'session' in bids_basename.entities
+    print(bids_basename.entities)
+    assert all(bids_basename.entities.get(entity) is None
+               for entity in ['task', 'run', 'recording', 'acquisition',
+                              'space', 'processing',
+                              'prefix', 'suffix'])
+
+    # test updating functionality
+    bids_basename.update(acquisition='03', run='2', session='02',
+                         task=None)
+    assert bids_basename.subject == subject_id
+    assert bids_basename.session == '02'
+    assert bids_basename.acquisition == '03'
+    assert bids_basename.run == '2'
+    assert bids_basename.task is None
+
+    new_bids_basename = bids_basename.copy().update(task='02',
+                                                    acquisition=None)
+    assert new_bids_basename.task == '02'
+    assert new_bids_basename.acquisition is None
+
+    # equality of bids basename
+    assert new_bids_basename != bids_basename
+    assert new_bids_basename == bids_basename.copy().update(task='02',
+                                                            acquisition=None)
+
+    # error check
+    with pytest.raises(ValueError, match='Key must be one of*'):
+        bids_basename.update(sub=subject_id, session=session_id)
+
+    # test repr
+    bids_path = make_bids_basename(subject='01', session='02',
+                                   task='03', suffix='ieeg.edf')
+    assert repr(bids_path) == 'BIDSPath(sub-01_ses-02_task-03_ieeg.edf)'